+++ packages/wagmi/src/utils/defaultWagmiReactConfig.ts
@@ -2,10 +2,9 @@ import '@web3modal/polyfills'
 
 import type { CreateConfigParameters, CreateConnectorFn, Config } from 'wagmi'
 import { createConfig } from 'wagmi'
-import { coinbaseWallet } from 'wagmi/connectors'
+import { coinbaseWallet, walletConnect } from 'wagmi/connectors'
 
 import { emailConnector } from '../connectors/EmailConnector.js'
-import { alphaWalletConnect } from '../connectors/alphaWalletConnect.js'
 import { getTransport } from './helpers.js'
 
 export type ConfigOptions = Partial<CreateConfigParameters> & {
@@ -48,7 +47,7 @@ export function defaultWagmiConfig({
 
   // Enabled by default
   if (enableWalletConnect !== false) {
-    connectors.push(alphaWalletConnect({ projectId, metadata, showQrModal: false }))
+    connectors.push(walletConnect({ projectId, metadata, showQrModal: false }))
   }
 
   if (enableCoinbase !== false) {